CurrentScene 가져오는 방법 SceneManager.GetActiveScene()으로 통일
- 현재 currentScene을 가져오는 방법은 GameManager.GetVariable("CurrentScene")과 SceneManager.GetActiveScene() 두가지가 있음
- 그런데 GameManager.GetVariable("CurrentScene")은 현재 씬이 아니라 이전에 저장된 씬으로 사용하는것이 맞음
- 이전에 저장된 씬을 가져오는 경우가 아닌, 현재 열려있는 씬(또는 인덱스)을 가져와야 하는 경우 SceneManager.GetActiveScene()으로 통일하도록 변경하였음
- 병합 후에는 variables 시트에서 CurrentScene을 SavedScene으로 변경하여야함

+++ Assets/1. Scripts/Managers/UI Manager.cs
@@ -428,8 +428,8 @@ public class UIManager : MonoBehaviour {
     
     public void OnLeftButtonClick() {
         switch (SceneManager.Instance.GetActiveScene()) {
-            case (int)SceneType.ROOM_1:
-            case (int)SceneType.ROOM_2:
+            case SceneType.ROOM_1:
+            case SceneType.ROOM_2:
                 RoomManager.Instance.MoveSides(-1);
                 break;
         }
@@ -437,8 +437,8 @@ public class UIManager : MonoBehaviour {
     
     public void OnRightButtonClick() {
         switch (SceneManager.Instance.GetActiveScene()) {
-            case (int)SceneType.ROOM_1:
-            case (int)SceneType.ROOM_2:
+            case SceneType.ROOM_1:
+            case SceneType.ROOM_2:
                 RoomManager.Instance.MoveSides(1);
                 break;
         }
@@ -449,13 +449,13 @@ public class UIManager : MonoBehaviour {
             MemoManager.Instance.OnExit();
         else
             switch (SceneManager.Instance.GetActiveScene()) {
-                case (int)SceneType.START:
+                case SceneType.START:
                     SetUI(eUIGameObjectName.Album, false, true, FloatDirection.Down);
                     SetUI(eUIGameObjectName.AlbumButton, true);
                     SetUI(eUIGameObjectName.ExitButton, false);
                     break;
-                case (int)SceneType.ROOM_1:
-                case (int)SceneType.ROOM_2:
+                case SceneType.ROOM_1:
+                case SceneType.ROOM_2:
                     RoomManager.Instance.OnExitButtonClick();
                     break;
                 default:
@@ -632,7 +632,7 @@ public class UIManager : MonoBehaviour {
     // method to switch mouse cursor
     public void SetCursorAuto() {
         bool isDefault = GameManager.Instance.GetIsBusy()
-                         || SceneManager.Instance.GetActiveScene() is (int)SceneType.START or (int)SceneType.ENDING
+                         || SceneManager.Instance.GetActiveScene() is SceneType.START or SceneType.ENDING
                          || (RoomManager.Instance && RoomManager.Instance.imageAndLockPanelManager.isLockObjectActive)
                          || isCursorTouchingUI;
         ChangeCursor(isDefault);
